factor out common select columns functionality to base component: candidate-source-base.ts
